requirements.txt
streamlit
Groq
python-dotenv